+++ packages/components/src/Image/src/SvgImage.module.css
