+++ packages/components/src/Image/src/SvgImageContext.ts
@@ -1,7 +1,7 @@
 import { createContext } from "react";
 import type { ContextValue } from "react-aria-components";
 
-import type { SvgImageProps } from "./svg-image.tsx";
+import type { SvgImageProps } from "./SvgImage.tsx";
 
 export const SvgImageContext = createContext<ContextValue<SvgImageProps, SVGSVGElement>>({});
