+++ packages/components/src/Image/src/index.ts
@@ -1,5 +1,5 @@
 export * from "./Image.tsx";
 export * from "./ImageContext.ts";
-export * from "./svg-image-context.ts";
-export * from "./svg-image.tsx";
+export * from "./SvgImage.tsx";
+export * from "./SvgImageContext.ts";
